+++ src/excel2sql/__init__.py
@@ -9,5 +9,5 @@
 """
 from __future__ import annotations
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'
 __all__ = ['__version__']